+++ backend/database_migration/tables/tblprogress_v1.py
@@ -4,7 +4,7 @@ def create_table(conn):
     with conn.cursor() as cur:
         cur.execute(f"""
             CREATE TABLE {TABLE_NAME} (
-                pg1_id bigint primary key,
+                pg1_id bigserial primary key,
                 pg1_created_at timestamptz default now(),
                 pg1_song_struct_id bigint references song_structure_tbl(id),
                 pg1_style text,